+++ CHANGES.md
@@ -5,6 +5,7 @@
 - MongoDB: Rename columns with leading underscores to use double leading underscores
 - MongoDB: Add support for UUID types
 - MongoDB: Improve reading timestamps in previous BSON formats
+- MongoDB: Fix processing empty arrays/lists. By default, assume `TEXT` as inner type.
 
 ## 2024/09/02 v0.0.21
 - DynamoDB: Add special decoding for varied lists.